requirements.txt
pandas
numpy
seaborn
matplotlib
scikit-learn
sqlalchemy
jupyter
tidyverse
caret
ggplot
python-dotenv
reportlab
psycopg2